댓글 달기, 메인화면 aside(추후 추천 또는 confidence로 top10 글 불러올 예정), 글 볼때 page 유지

+++ src/main/java/com/test/postgresql_test/domain/Entity/Reply.java
@@ -22,7 +22,7 @@ public class Reply {
     @Lob
     private String comment;
 
-    private Long level;
+    private int level;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JsonBackReference
